requirements.txt
Django==3.2.3
psycopg2-binary==2.8.6
djangorestframework==3.12.4
markdown
django-filter==2.4.0
drf_yasg
django-rest-swagger
